refactor: 수익금 클래스 개선
- 수익금을 Double 타입으로 저장합니다.

+++ src/main/kotlin/blackjack/domain/betting/ProfitAmount.kt
@@ -1,4 +1,4 @@
 package blackjack.domain.betting
 
 @JvmInline
-value class ProfitAmount(val value: Int)
+value class ProfitAmount(val value: Double)
